backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
reportlab==4.0.7
openpyxl==3.1.2
xlrd==2.0.1
boto3==1.34.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
websockets==12.0
firebase-admin==6.2.0
openai==1.3.7 